refactored virtualbox provider to vagrant
