+++ lib/mccloud.rb
@@ -13,7 +13,7 @@ module Mccloud
 end
 
 # # Default I18n to load the en locale
-I18n.load_path << File.expand_path("templates/locales/en.yml", Mccloud.source_root)
+I18n.load_path << File.expand_path("lib/mccloud/templates/locales/en.yml", Mccloud.source_root)
 
 # Load the things which must be loaded before anything else
 require 'mccloud/cli'
