--- lib/mccloud/provider/virtualbox/vm/up.rb
@@ -1,11 +1,0 @@
-module Mccloud::Provider
-  module Virtualbox
-    module VmCommand
-
-        def up(command,options={})
-            Vagrant::CLI.start(["up"],:env => @provider.raw)
-        end
- 
-    end #module
-  end #module
-end #module
